// src/app/shared/modules/dynamic-form/models/field-config.interface.ts
import { ValidatorFn } from '@angular/forms';

export interface FieldConfig {
  disabled?: boolean;
  label?: string;
  name: string;
  options?: string[];
  placeholder?: string;
  type: string;
  validation?: ValidatorFn[];
  value?: any;
}
export interface MyOptions {
  value: string;
  text: string;
}